requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
confluent-kafka==2.3.0
dapr==1.12.0
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
